test(integration): add integration test suite for core services
Add comprehensive integration tests that run against real WordPress:
- EnrollmentService: enrollment, cancellation, status checks
- VoucherService: validation, discount calculations, expiry/usage
- QuoteUpdateHandler: auth, access control, quote updates
- ProfileHandler: profile updates via API
- ApiEndpoint: filter registration tests

Bug fixes discovered during testing:
- Add missing RegistrationStatus::Completed enum case
- Fix VoucherService.hydrateVoucher to strip _ntdst_ prefix from meta
- Fix EditionService.onRegistrationCancelled to cast edition_id to int

+++ web/app/mu-plugins/stride-core/Modules/Edition/EditionService.php
@@ -187,7 +187,7 @@ final class EditionService extends AbstractService implements EditionQueryInterf
      */
     public function onRegistrationCancelled(array $data): void
     {
-        $editionId = $data['edition_id'] ?? 0;
+        $editionId = (int) ($data['edition_id'] ?? 0);
         $currentStatus = $this->getStatus($editionId);
 
         if ($editionId && $currentStatus === EditionStatus::Full) {
